+++ src/components/Modal/Modal.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactModal from "react-modal";
-import { css } from "@emotion/core";
 import { Tab, Tabs, TabList, TabPanel } from "react-tabs";
+import { modalStyle, tabsStyle} from "./assets/style";
 
 interface IProps {
   children?: React.ReactNode;
@@ -13,39 +13,6 @@ interface IProps {
 }
 const isTest = process.env.NODE_ENV === "test";
 !isTest && ReactModal.setAppElement("#root");
-
-const modalStyle = {
-  content: {
-    backgroundColor: "var(--color-background)",
-    border: 0,
-    color: "var(--color-text)",
-    left: "50%",
-    opacity: 1,
-    overflowX: "hidden",
-    padding: "2rem 2rem",
-    textAlign: "center",
-    top: "50%",
-    transform: "translate(-50%, -50%)",
-    minWidth: "20rem",
-    minHeight: "22.5rem",
-    zIndex: 1000
-  },
-  overlay: {
-    backgroundColor: "#0000007F",
-    position: "absolute",
-    height: "37.5rem",
-    width: "50rem",
-    zIndex: 999
-  }
-};
-
-const tabsStyle = css`
-  & > .react-tabs__tab-list {
-    position: absolute;
-    top: 0;
-    left: 0;
-  }
-`;
 
 const Modal: React.FunctionComponent<IProps> = ({
   children,
