+++ src/components/Modal/TableSelect.tsx
@@ -10,58 +10,13 @@ import {
   setUserSeat
 } from "../../store/actions";
 import { IState } from "../../store/initialState";
-import ModalButtonsWrapper from "./ModalButtonsWrapper";
+// import ModalButtonsWrapper from "./ModalButtonsWrapper";
+import { Table, TableArea , ModalButtonsWrapper} from "./assets/style";
 
 interface IProps {
   dispatch: (arg: object) => void;
   state: IState;
 }
-
-const Table = styled.table`
-  border-spacing: 0;
-  color: white;
-  cursor: pointer;
-  font-size: var(--font-size-s);
-  text-align: center;
-  width: 100%;
-
-  tbody {
-    padding: 0.5rem;
-
-    & tr.selected {
-      background-color: var(--color-primaryLight);
-      color: var(--color-background);
-      cursor: default;
-    }
-
-    & tr:hover:not(.selected) {
-      background-color: var(--color-accent);
-      color: var(--color-background);
-    }
-  }
-
-  td {
-    font-family: var(--font-family-secondary);
-    font-weight: 400;
-    padding: 0.25rem;
-  }
-
-  thead {
-    & tr {
-      background-color: var(--color-background);
-
-      & th {
-        padding-bottom: 0.25rem;
-      }
-    }
-  }
-`;
-
-const TableArea = styled.div`
-  background-color: var(--color-darkGray);
-  height: 14rem;
-  overflow: scroll;
-`;
 
 const TableSelect: React.FunctionComponent<IProps> = ({ dispatch, state }) => {
   const [tableList, setTableList] = useState([]);
